Fix incorrect scylla-server sysconfig file edit for i3 memflush controller
The previous version that I sent was wrong - and I still don't
understand how is it that it didn't failed in some tests. Adding to the
puzzle, when building a AMI with this code and the 1.7 branch with 2.0
AMIs built for will just complain that the parameters themselves are
invalid, but will parse them right.

But 2.0 AMIs won't work: systemd does not expand variables in the
sysconfig files, and just treat them as simple assignment. Therefore the
syntax I have used here, SCYLLA_ARGS="$SCYLLA_ARGS..." will get the 2.0
to complains that $SCYLLA_ARGS is not a valid parameter for Scylla.

Systemd's documentation confirms that expansions shouldn't happen, so
let us not insist: as much as I hate the sed route, let's use that
instead.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20170726185222.2100-1-glauber@scylladb.com>