Move to kernel-ml kernel stream
We have to change kernel for our AMI one way or another, because the 2017.03
stream from Amazon does not seem to include the meltdown fixes. If we are going
to do it, we might as well move to the Elrepo stream, which ships mainline
kernels to CentOS.

In 2018, mainline kernels are much more trustworthy than they used to be.
Ubuntu follows them quite closely, and even the kernels that Amazon ships are
pretty close to the mainline tree.

I have boot tested an AMI with this kernel, and I verified that the ENA driver
that this kernel ships is at the same version as the one in the Amazon tree for
the 2017.03 stream.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180109231103.2601-1-glauber@scylladb.com>