change default roles for EBS / ephemeral
A while ago we have introduced support for using EBS in some form
in the AMI. This came as a need from people doing commitlog in SSD
disks and data in EBS disks.

To support data, we have introduced roles for those disks, and set
the default roles according to that use case.

However, as we are now seeing in the field, it is a lot more common
for uses to attach EBS devices for ad-hoc extraneous operations, like
backing up snapshots.

Once the AMI detects the EBS there, it will try to mount it as data.
Let's make the default what's actually common, with ephemeral being
used for everything and EBS being unused.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180503171913.32677-1-glauber@scylladb.com>