don't try to add busy devices to the RAID.
For old instances, we find the list of devices that we would like to add
to our RAID from the instance metadata.

NVMe-enabled instances break that because information about those
devices are not added to the instance metadata at all. What we have been
doing for them is to just scan /dev/ for all NVMe devices and add them
to the list of ephemeral devices. We could do that, because the root
device was always an EBS device, so whatever else we found had to be an
ephemeral drive.

This logic breaks with i3.metal. Their root devices are exposed as NVMe
devices. We have to do something else.

This patch will scan the system and create a list of devices that are
busy at the time our script runs. Busy devices are listed as mount
points, or are slave-parts of arrays that are themselves listed as mount
points.

With this patch we can create the array successfully on i3.metal
instances.

Signed-off-by: Glauber Costa <glauber@scylladb.com>
Message-Id: <20180611171112.27482-1-glauber@scylladb.com>