scylla_create_devices: fix mouting RAID volume after reboot
Since we use this script to mount RAID volume that previously
constructed, we need to pass NVMe devices that belongs to RAID.

Currently we use device_is_busy() to exclude NVMe device that used for root
volume, but it mistakenly drop devices that belongs to RAID too.
Result of that the script has no devices to use for RAID.

We need to exclude only root volume instead.

Ref scyllad/scylla#3640

Signed-off-by: Takuya ASADA <syuu@scylladb.com>
Message-Id: <20180810021426.28163-1-syuu@scylladb.com>