package scylla-ami as rpm
Now scylla-ami is not submodule of scylla repo, it will works as
independent repository just like scylla-jmx and scylla-tools,
provides .rpm package to install AMI scripts on AMI.

We dropped scylla_install_ami since we copied it to scylla repo,
since since it requires to install scylla .rpms, cannot pacakge in
scylla-ami rpm.

On new scylla-ami .rpm, we requires to packages 'scylla-ixgbevf' and
'scylla-ena' DKMS .rpm.
It will automatically build kernel modules for current kernel.

A repo of the driver packages is on
https://copr.fedorainfracloud.org/coprs/scylladb/scylla-ami-drivers/

Signed-off-by: Takuya ASADA <syuu@scylladb.com>
Message-Id: <20180821201013.4543-1-syuu@scylladb.com>