ami: scylla_ami_setup: call sysconfig_setup after create_devices
We need to call sysconfig_setup after create_devices because the former
depends on the result of the later, in particular '/var/lib/scylla' needs
to be mounted at the correct device (e.g. NVMe-based RAID0) in order to
allow perftune.py to detect the configuration mode correctly.

Fixes sylladb/scylla#3911

Signed-off-by: Vlad Zolotarov <vladz@scylladb.com>
Message-Id: <20181115234459.27604-1-vladz@scylladb.com>